Clear the compiler warnings and keep them cleared
70 unique warnings, not the 187 I quoted earlier -- that figure counted the same
header warning once per translation unit.

Twelve were substantive, and two were bugs rather than noise:

  Logger's timestamp used std::localtime, which returns a pointer into one
  shared static buffer. Logger::log holds a mutex, but every subsystem
  constructs its own Logger with its own mutex and resource loading runs on a
  worker, so the mutex does not cover it. Now localtime_s / localtime_r.

  FrameGraphJson used strerror, the same hazard, while building an error
  message. Now std::generic_category().message().

The rest:

  ApplicationBase narrowed the swapchain image count from size_t without
  saying so; SceneAPI discarded a [[nodiscard]] deliberately and now says so;
  IBLGenerator computed Lx and Ly every iteration of the BRDF integration and
  used neither, because the normal is fixed at (0,0,1) so only Lz is read;
  FrameGraphExport had a second potentialSavings local that was always zero
  beside the real one; EntityRenderExecutor shadowed an outer iterator and
  computed an isSkinnedPass flag for a branch nobody ever wrote; four builder
  parameters shadowed the members they assigned; and Scene::saveToFile /
  loadFromFile caught an exception, discarded the reason and returned false,
  which made a full disk indistinguishable from a malformed file -- they now
  report what went wrong.

getenv in AssetPaths keeps its warning suppressed locally rather than fixed:
_dupenv_s is MSVC-only, and the hazard it warns about does not apply when the
value is copied immediately.

The remaining 59 were unreferenced parameters, mostly virtual hooks that
deliberately ignore an argument. Their names are commented rather than deleted,
so the documentation the name provides survives. Applied from the compiler's own
C4100 output rather than by matching source, so only parameters MSVC confirmed
unused were touched.

One test changed instead of being silenced: EntityBuilder.WithParent_MultipleChildren
named two children and then only counted them, so it would have passed if the
second build overwrote the first. It now asserts which children, in what order,
and that both point back at the parent.

SHOONYAKASHA_WERROR keeps it at zero. ON in CI, OFF by default -- a compiler
version that invents a new warning should fail the pipeline, not somebody's
first local build. Turning it on immediately failed on cgltf's fopen and strcpy,
so the vendored include directories are marked SYSTEM (with /external:W0 on
MSVC, which /W4 needs to honour it). Warnings from code we cannot fix without
forking it are not ours to treat as errors.

Library, examples and tests all build clean at /W4 and at /W4 /WX. 681 C++
tests, 22 Python tests, eight examples exit 0 with no validation messages.

+++ src/App/ApplicationBase.cpp
@@ -126,7 +126,8 @@ void ApplicationBase::initializeVulkan() {
     m_fontLoader = std::make_unique<FontLoader>(*m_device);
 
     m_commandManager = std::make_unique<VulkanCommandManager>(*m_device);
-    m_commandBuffers = m_commandManager->createCommandBuffers(m_swapChain->getImageCount());
+    m_commandBuffers = m_commandManager->createCommandBuffers(
+        static_cast<uint32_t>(m_swapChain->getImageCount()));
 
     m_logger->log(LogLevel::Info, "Vulkan systems initialized");
 }
